# old_api_spec.yaml
openapi: 3.0.0
info:
  title: AI Life Coach API
  version: 1.0.0
servers:
  - url: https://api.yourapp.com
  - url: http://localhost:8000
paths:
...
  /api/admin/panel/users/list:
    get:
      tags:
        - admin_panel
      summary: Retrieve a paginated list of users
      description: Returns a list of users with optional filters (country, user_id, email) and ordering (activity, signup_date, country, total_files, total_records).
      operationId: getUsersList
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
          description: The page number to retrieve (for pagination).
        - name: page_size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 10
          description: The number of users to return per page (for pagination).
        - name: country
          in: query
          required: false
          schema:
            type: string
          description: Filter users by country.
        - name: user_id
          in: query
          required: false
          schema:
            type: integer
          description: Filter by a specific user ID.
        - name: email
          in: query
          required: false
          schema:
            type: string
          description: Filter by email address (e.g., partial or exact match).
        - name: sort_by
          in: query
          required: false
          schema:
            type: string
            enum: [activity, signup_date, country, total_files, total_records]
          description: Order results by user activity, signup date, country, total_files, or total_records.
        - name: sort_order
          in: query
          required: false
          schema:
            type: string
            enum: [asc, desc]
          description: The sorting order (ascending or descending). Typically defaults to `desc` for dates.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Paginated list of users retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  page:
                    type: integer
                    example: 1
                  page_size:
                    type: integer
                    example: 10
                  total:
                    type: integer
                    description: Total number of users matching the filters.
                    example: 125
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        user_id:
                          type: integer
                          example: 101
                        user_email:
                          type: string
                          example: "user@example.com"
                        country:
                          type: string
                          example: "US"
                        attributes:
                          type: object
                          description: A JSON object representing user attributes or metadata.
                          example:
                            plan: "premium"
                            language: "en"
                        signup_date:
                          type: string
                          format: date-time
                          example: "2024-01-01T12:00:00Z"
                        last_activity_date:
                          type: string
                          format: date-time
                          example: "2024-01-10T09:15:00Z"
                        number_of_files:
                          type: integer
                          example: 5
                        list_of_file_ids:
                          type: array
                          items:
                            type: integer
                          example: [123, 124, 125]
                        number_of_records:
                          type: integer
                          example: 150
        '400':
          $ref: '#/components/responses/BadRequestError'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
...
  responses:
...
    BadRequestError:
      description: Bad request.
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
                example: Bad request
    InternalServerError:
      description: Internal server error.
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
                example: Internal server error

  # securitySchemes:
  #   bearerAuth:
  #     type: http
  #     scheme: bearer
  #     bearerFormat: JWT
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
      description: API key needed to authorize requests.












            
  # /journal/entry:
  #   post:
  #     summary: Submit a new journal entry
  #     tags:
  #       - journal
  #     requestBody:
  #       required: true
  #       content:
  #         application/json:
  #           schema:
  #             $ref: '#/components/schemas/JournalEntrySubmit'
  #     security:
  #       - bearerAuth: []
  #     responses:
  #       '201':
  #         description: Created
          
            
...